+++ changelog.md
@@ -10,6 +10,7 @@ Codion Change Log
 ### is.codion.swing.common.ui
 - DefaultFilterTableCellRenderer.DefaultBuilder bug fixed, useBooleanRenderer now initialized before settings, in order for a correct horizontal alignment for boolean columns.
 - FilterTableCellRenderer.Factory.create() T type parameter removed.
+- FilterTableColumnModel.identifiers() added.
 ### is.codion.framework.db
 - DefaultLocalEntityConnection.delete() bug fixed, database exception now with the correct Operation.
 ### is.codion.swing.framework.model 
@@ -31,6 +32,7 @@ Codion Change Log
 - EntityPanel.applyPreferences() now public.
 - EntityApplicationPanel now handles applying preferences for its entity panels.
 - EntityTablePanel now remembers and reuses the dependencies dialog size.
+- EntityTablePanel now remembers and reapplies column user preferences for dependency panels, related refactoring.
 
 ## 0.18.18
 ### is.codion.common.db